requirements.txt
google-adk
langchain
dotenv
langchain-openai
langchain-community
litellm
pydantic
python-dotenv
fastapi
uvicorn
requests
langchain-pinecone
pinecone-client
